use go testing framework
