+++ .github/workflows/test-act.yml
@@ -53,7 +53,7 @@ jobs:
 
       - name: Run act tests
         run: |
-          go run main.go
+          go test -v
         working-directory: tests/act
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
