+++ tests/act/go.mod
@@ -3,3 +3,10 @@ module github.com/grafana/plugin-ci-workflows/tests/act
 go 1.25.4
 
 require github.com/google/uuid v1.6.0
+
+require (
+	github.com/davecgh/go-spew v1.1.1 // indirect
+	github.com/pmezard/go-difflib v1.0.0 // indirect
+	github.com/stretchr/testify v1.11.1
+	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
